fix/file-player-external-packages
