+++ packages/ui/src/components/features/file-players/sheet/hooks/use-sheet-data.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import * as XLSX from 'xlsx';
+import { parseSheetFile } from '@/server-actions/sheet-parser';
 
 type SheetData = (string | number | boolean | null)[][];
 
@@ -16,18 +16,7 @@ export function useSheetData(file: File) {
 		async function parse() {
 			try {
 				const buffer = await file.arrayBuffer();
-
-				const workbook = XLSX.read(buffer, {
-					type: 'array',
-				});
-
-				const firstSheetName = workbook.SheetNames[0];
-				const sheet = workbook.Sheets[firstSheetName];
-
-				const json = XLSX.utils.sheet_to_json(sheet, {
-					header: 1,
-					defval: '',
-				}) as SheetData;
+				const json = await parseSheetFile(buffer);
 
 				if (!cancelled) setData(json);
 			} catch (err) {
